react-query for rankings page
